# Store elements in screenLine